+++ lib/util/color_filter_component.dart
@@ -1,6 +1,5 @@
 import 'dart:ui';
 
-import 'package:bonfire/base/bonfire_game.dart';
 import 'package:bonfire/base/game_component.dart';
 import 'package:bonfire/util/game_color_filter.dart';
 import 'package:bonfire/util/priority_layer.dart';
@@ -33,12 +32,6 @@ class ColorFilterComponent extends GameComponent
       );
       canvas.restore();
     }
-  }
-
-  @override
-  set gameRef(BonfireGame gameRef) {
-    colorFilter.gameRef = gameRef;
-    super.gameRef = gameRef;
   }
 
   @override