+++ lib/util/text_damage_component.dart
@@ -1,6 +1,5 @@
 import 'dart:math';
 
-import 'package:bonfire/base/bonfire_game.dart';
 import 'package:bonfire/util/priority_layer.dart';
 import 'package:flame/components.dart';
 import 'package:flame/extensions.dart';
@@ -9,8 +8,7 @@ import 'bonfire_game_ref.dart';
 
 enum DirectionTextDamage { LEFT, RIGHT, RANDOM, NONE }
 
-class TextDamageComponent extends TextComponent
-    with BonfireHasGameRef<BonfireGame> {
+class TextDamageComponent extends TextComponent with BonfireHasGameRef {
   final String text;
   final DirectionTextDamage direction;
   final double maxDownSize;