+++ CHANGELOG.md
@@ -1,7 +1,11 @@
-# [next]
+# 1.9.11
 - Fix [#203](https://github.com/RafaelBarbosatec/bonfire/issues/203) - Web build with late initialisation on animations
 - Create interfaces to facility access methods of the `ColorFilter` and `Lighting`.
-- Add `replaceAnimation` method in `DirectionAnimation`
+- Add `replaceAnimation` method in `DirectionAnimation`. now you can use this method in SimplePlayer or Enemy to replace `SimpleDirectionAnimation`.
+- Now gameRef is `BonfireGameInterface`.
+- BREAKING CHANGE:
+    - rename `gameRef.changeJoystickTarget` to `gameRef.addJoystickObserver`
+    - remove `gameRef.addComponent`. now use `gameRef.add` or `gameRef.addAll`
 
 # 1.9.10
 - Fix problem render big tilesets [#200](https://github.com/RafaelBarbosatec/bonfire/issues/200).